+++ code/game/g_weapon.c
@@ -1589,7 +1589,7 @@ static void WP_FireRocket( gentity_t *ent, qboolean altFire )
 	int	damage	= ROCKET_DAMAGE;
 	int	vel = ROCKET_VELOCITY;
 	int dif = 0;
-	float rTime;
+	int rTime;
 	gentity_t *missile;
 
 	if ( altFire )
